Don't bundle-build server
