+++ packages/build-tooling/package.json
@@ -5,6 +5,8 @@
   "main": "index.mjs",
   "type": "module",
   "bin": {
+    "chokidar": "./node_modules/.bin/chokidar",
+    "concurrently": "./node_modules/.bin/concurrently",
     "move-dts": "./move-dts.mjs",
     "pkg": "./node_modules/.bin/pkg",
     "swc": "./node_modules/.bin/swc",
@@ -22,6 +24,8 @@
     "@swc/cli": "^0.1.62",
     "@swc/core": "^1.3.37",
     "babel-loader": "^9.1.0",
+    "chokidar-cli": "^3.0.0",
+    "concurrently": "^8.2.1",
     "glob": "^10.3.3",
     "nodemon-webpack-plugin": "^4.8.1",
     "pkg": "^4.4.8",
