+++ packages/sync-server/package.json
@@ -19,15 +19,19 @@
     "node": "^16.16.0"
   },
   "scripts": {
-    "start": "node ./dist/app.bundle.js",
-    "setup-dev": "node ./dist/app.bundle.js setup",
-    "start-dev": "rimraf ./dist/ && webpack --config webpack.config.dev.mjs",
+    "build": "swc --delete-dir-on-start --out-dir dist --copy-files --source-maps true app",
+    "start": "node dist",
+
     "test": "NODE_ENV=test jest",
-    "test-watch": "yarn test --watch",
-    "build": "rimraf ./dist/ && webpack --config webpack.config.prod.mjs",
-    "build-dev": "rimraf ./dist/ && NOWATCH=true webpack --config webpack.config.dev.mjs",
     "lint": "eslint app",
-    "shell": "yarn build-dev && node ./dist/app.bundle.js shell"
+
+    "setup-dev": "node dist setup",
+    "shell": "yarn build && node dist shell",
+
+    "start-watch": "chokidar \"dist/**/*.js\" -c \"yarn run start\"",
+    "test-watch": "yarn test --watch",
+    "build-watch": "yarn run build --watch",
+    "start-dev": "concurrently npm:build-watch npm:start-watch"
   },
   "devDependencies": {
     "@beyondessential/eslint-config-js": "^1.1.0",
@@ -36,7 +40,6 @@
     "ajv": "^8.6.3",
     "chai": "^4.1.2",
     "eslint": "^5.15.2",
-    "eslint-import-resolver-webpack": "^0.11.0",
     "eslint-plugin-import": "^2.22.1",
     "jest": "^29.3.1",
     "jest-expect-message": "^1.1.3",
